CloudFunction/requirements.txt
tensorflow==2.4.1
numpy==1.16.4
requests==2.22.0
google-cloud-storage==1.17.0